pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "SEO Page Analyzer is a web tool to check a web page readiness for SEO (Search Engine Optimization)"
authors = ["Alexandr Sakhatov <sakhatoff@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]
repository = "https://github.com/Unt0ten/python-project-83"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.3"
gunicorn = "20.1.0"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
validators = "^0.22.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
